+++ hammer-sqldb/src/main/java/cn/featherfly/hammer/sqldb/dsl/entity/condition/AbstractMulitiEntityPropertyExpression.java
@@ -191,80 +191,126 @@ public abstract class AbstractMulitiEntityPropertyExpression<E, C extends Condit
      * @return the property mapping
      */
     public PropertyMapping<?> getPropertyMapping(Object value) {
-        int index = this.index.get();
         if (propertyList.size() == 1) {
-            ClassMapping<?, JdbcPropertyMapping> classMapping = expression.getClassMapping(index);
+            ClassMapping<?, JdbcPropertyMapping> classMapping = expression.getClassMapping(index.get());
             return classMapping.getPropertyMapping(LambdaUtils.getLambdaPropertyName(propertyList.get(0)));
-        } else if (propertyList.size() == 2) {
-            ClassMapping<?, JdbcPropertyMapping> classMapping = expression.getClassMapping(index);
-            JdbcPropertyMapping pm = classMapping
-                .getPropertyMapping(LambdaUtils.getLambdaPropertyName(propertyList.get(0)));
-            if (value == null) {
-                // ENHANCE 这个查询值为null则直接返回对象映射的逻辑后续考虑是否合理
-                return pm;
+        } else if (propertyList.size() > 1) {
+            PropertyMapping<?> pm = null;
+            for (int i = 0; i < propertyList.size() - 1; i++) {
+                pm = getSubPropertyMapping(propertyList.get(i), propertyList.get(i + 1), value);
             }
-
-            SerializedLambdaInfo propertyInfo = LambdaUtils.getLambdaInfo(propertyList.get(1));
-            String pn = propertyInfo.getPropertyName();
-            if (pm.getMode() == Mode.EMBEDDED) {
-                JdbcPropertyMapping spm = pm.getPropertyMapping(pn);
-                if (spm == null) {
-                    throw new SqldbHammerException(Str.format("no property mapping found for {0}.{1}.{2}",
-                        classMapping.getType().getSimpleName(), pm.getPropertyFullName(), pn));
-                }
-                return spm;
-            } else if (Mode.MANY_TO_ONE == pm.getMode()) {
-                // } else if (Mode.MANY_TO_ONE == pm.getMode() || Mode.ONE_TO_ONE == pm.getMode()) {
-                // YUFEI_TEST 需要测试ONE_TO_ONE
-                JdbcPropertyMapping spm = pm.getPropertyMapping(pn);
-                if (spm != null) {
-                    return spm;
-                } else {
-                    @SuppressWarnings("unchecked")
-                    JdbcClassMapping<E> cm = factory.getClassMapping((Class<E>) pm.getPropertyType());
-                    // 这里需要join，在条件设置中需要判断对象是否已经join，如果没有join，则需要在设置查询参数时，先join
-                    // 所以需要在条件中记录已经join的对象关系来判断是否需要join
-                    spm = cm.getPropertyMapping(pn);
-                    if (spm != null) {
-                        queryRelation.join(Join.LEFT_JOIN, index, pm.getPropertyName(), cm, false);
-                        this.index.incrementAndGet();
-                        return spm;
-                    } else {
-                        throw new SqldbHammerException(
-                            Str.format("no property mapping found for {0}.{1}", cm.getType().getSimpleName(), pn));
-                    }
-                }
-            } else if (pm.getMode() == Mode.ONE_TO_MANY) {
-                // IMPLSOON 未实现一对多
-                // 在条件设置中需要判断对象是否已经join，如果没有join，则需要在设置查询参数时，先join
-                // 所以需要在条件中记录已经join的对象关系来判断是否需要join
-                throw new NotImplementedException();
-            } else if (pm.getMode() == Mode.SINGLE) {
-                // YUFEI_TEST 待测试，propertyList.size() > 1 时 propertyList.get(0)就不应该是Mode.SINGLE，所以应该永远进不了这个逻辑
-                throw new NotImplementedException();
-            }
-            throw new UnsupportedException();
+            return pm;
         } else {
-            // IMPLSOON 未实现，出现多次对象嵌套，只可能出现在关联对象的情况即，即多次获取ManyToOne(OneToOne)的属性
-            // 例如： order.getOwner().getUserInfo().getAddress().getNo()  -  order > user > user_info > address.no
             throw new NotImplementedException();
         }
+        //        else if (propertyList.size() == 2) {
+        //            ClassMapping<?, JdbcPropertyMapping> classMapping = expression.getClassMapping(index);
+        //            JdbcPropertyMapping pm = classMapping
+        //                .getPropertyMapping(LambdaUtils.getLambdaPropertyName(propertyList.get(0)));
+        //            if (value == null) {
+        //                // ENHANCE 这个查询值为null则直接返回对象映射的逻辑后续考虑是否合理
+        //                return pm;
+        //            }
+        //
+        //            SerializedLambdaInfo propertyInfo = LambdaUtils.getLambdaInfo(propertyList.get(1));
+        //            String pn = propertyInfo.getPropertyName();
+        //            if (pm.getMode() == Mode.EMBEDDED) {
+        //                JdbcPropertyMapping spm = pm.getPropertyMapping(pn);
+        //                if (spm == null) {
+        //                    throw new SqldbHammerException(Str.format("no property mapping found for {0}.{1}.{2}",
+        //                        classMapping.getType().getSimpleName(), pm.getPropertyFullName(), pn));
+        //                }
+        //                return spm;
+        //            } else if (Mode.MANY_TO_ONE == pm.getMode()) {
+        //                // } else if (Mode.MANY_TO_ONE == pm.getMode() || Mode.ONE_TO_ONE == pm.getMode()) {
+        //                // YUFEI_TEST 需要测试ONE_TO_ONE
+        //                JdbcPropertyMapping spm = pm.getPropertyMapping(pn);
+        //                if (spm != null) {
+        //                    return spm;
+        //                } else {
+        //                    @SuppressWarnings("unchecked")
+        //                    JdbcClassMapping<E> cm = factory.getClassMapping((Class<E>) pm.getPropertyType());
+        //                    // 这里需要join，在条件设置中需要判断对象是否已经join，如果没有join，则需要在设置查询参数时，先join
+        //                    // 所以需要在条件中记录已经join的对象关系来判断是否需要join
+        //                    spm = cm.getPropertyMapping(pn);
+        //                    if (spm != null) {
+        //                        queryRelation.join(Join.LEFT_JOIN, index, pm.getPropertyName(), cm, false);
+        //                        this.index.incrementAndGet();
+        //                        return spm;
+        //                    } else {
+        //                        throw new SqldbHammerException(
+        //                            Str.format("no property mapping found for {0}.{1}", cm.getType().getSimpleName(), pn));
+        //                    }
+        //                }
+        //            } else if (pm.getMode() == Mode.ONE_TO_MANY) {
+        //                // IMPLSOON 未实现一对多
+        //                // 在条件设置中需要判断对象是否已经join，如果没有join，则需要在设置查询参数时，先join
+        //                // 所以需要在条件中记录已经join的对象关系来判断是否需要join
+        //                throw new NotImplementedException();
+        //            } else if (pm.getMode() == Mode.SINGLE) {
+        //                // YUFEI_TEST 待测试，propertyList.size() > 1 时 propertyList.get(0)就不应该是Mode.SINGLE，所以应该永远进不了这个逻辑
+        //                throw new NotImplementedException();
+        //            }
+        //            throw new UnsupportedException();
+        //        } else {
+        //            // IMPLSOON 未实现，出现多次对象嵌套，只可能出现在关联对象的情况即，即多次获取ManyToOne(OneToOne)的属性
+        //            // 例如： order.getOwner().getUserInfo().getAddress().getNo()  -  order > user > user_info > address.no
+        //            throw new NotImplementedException();
+        //        }
     }
 
-    //    protected ColumnElement getColumnElement(Object value) {
-    //        if (column != null && arithmeticColumnElement.get() != null) {
-    //            //            Dialect dialect, ColumnElement column, Object value,
-    //            //            ComparisonOperator comparisonOperator, String tableAlias, Predicate<?> ignoreStrategy
-    //            return new CompositeConditionColumnElement(expression.getJdbc().getDialect(), column, value,
-    //                arithmeticColumnElement.get(), expression.getAlias(index.get()), null);
-    //        } else if (column != null) {
-    //            return column;
-    //        } else if (arithmeticColumnElement.get() != null) {
-    //            return arithmeticColumnElement.get();
-    //        } else {
-    //            return null;
-    //        }
-    //    }
+    private PropertyMapping<?> getSubPropertyMapping(Serializable property, Serializable subProperty, Object value) {
+        final int currentIndex = index.get();
+        ClassMapping<?, JdbcPropertyMapping> classMapping = expression.getClassMapping(currentIndex);
+        JdbcPropertyMapping pm = classMapping
+            .getPropertyMapping(LambdaUtils.getLambdaPropertyName(property));
+        if (value == null) {
+            // ENHANCE 这个查询值为null则直接返回对象映射的逻辑后续考虑是否合理
+            return pm;
+        }
+
+        SerializedLambdaInfo propertyInfo = LambdaUtils.getLambdaInfo(subProperty);
+        String pn = propertyInfo.getPropertyName();
+
+        if (pm.getMode() == Mode.EMBEDDED) {
+            JdbcPropertyMapping spm = pm.getPropertyMapping(pn);
+            if (spm == null) {
+                throw new SqldbHammerException(Str.format("no property mapping found for {0}.{1}.{2}",
+                    classMapping.getType().getSimpleName(), pm.getPropertyFullName(), pn));
+            }
+            return spm;
+        } else if (Mode.MANY_TO_ONE == pm.getMode()) {
+            // } else if (Mode.MANY_TO_ONE == pm.getMode() || Mode.ONE_TO_ONE == pm.getMode()) {
+            // YUFEI_TEST 需要测试ONE_TO_ONE
+            JdbcPropertyMapping spm = pm.getPropertyMapping(pn);
+            if (spm != null) {
+                return spm;
+            } else {
+                @SuppressWarnings("unchecked")
+                JdbcClassMapping<E> cm = factory.getClassMapping((Class<E>) pm.getPropertyType());
+                // 这里需要join，在条件设置中需要判断对象是否已经join，如果没有join，则需要在设置查询参数时，先join
+                // 所以需要在条件中记录已经join的对象关系来判断是否需要join
+                spm = cm.getPropertyMapping(pn);
+                if (spm != null) {
+                    queryRelation.join(Join.LEFT_JOIN, currentIndex, pm.getPropertyName(), cm, false);
+                    index.incrementAndGet();
+                    return spm;
+                } else {
+                    throw new SqldbHammerException(
+                        Str.format("no property mapping found for {0}.{1}", cm.getType().getSimpleName(), pn));
+                }
+            }
+        } else if (pm.getMode() == Mode.ONE_TO_MANY) {
+            // IMPLSOON 未实现一对多
+            // 在条件设置中需要判断对象是否已经join，如果没有join，则需要在设置查询参数时，先join
+            // 所以需要在条件中记录已经join的对象关系来判断是否需要join
+            throw new NotImplementedException();
+        } else if (pm.getMode() == Mode.SINGLE) {
+            // YUFEI_TEST 待测试，propertyList.size() > 1 时 propertyList.get(0)就不应该是Mode.SINGLE，所以应该永远进不了这个逻辑
+            throw new NotImplementedException();
+        }
+        throw new UnsupportedException();
+    }
 
     protected ColumnElement getColumnElement(PropertyMapping<?> propertyMapping) {
         if (arithmeticColumnElement.get() != null) {